recomendationApp/requirements.txt
uvicorn
pandas
numpy
scikit-learn
python-dotenv
fastapi
requests
